Display time being played and remaining in player
